requirements.txt
pyautogui~=0.9.54
PyGetWindow~=0.0.9
pytesseract~=0.3.10
pillow~=10.3.0